+++ src/app/clusters/boolean-state-server/BUILD.gn
@@ -16,8 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("boolean-state-server") {
   sources = [
-    "boolean-state-cluster.cpp",
-    "boolean-state-cluster.h",
+    "BooleanStateCluster.cpp",
+    "BooleanStateCluster.h",
   ]
 
   public_deps = [
